requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
fastf1==3.0.0
filetype==1.2.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
frozenlist==1.8.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.32.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-groq==1.0.0
langchain-text-splitters==1.0.0
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-cli==0.4.4
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langgraph-supervisor==0.0.29
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
multidict==6.7.0
mypy_extensions==1.1.0
neon-api==0.3.0
numpy==1.26.4
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
referencing==0.37.0
requests==2.32.5
requests-cache==1.2.1
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
tenacity==9.1.2
thefuzz==0.22.1
timple==0.1.8
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
